+++ RMS/src/main/java/com/cg/RMS/repository/CompanyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 import com.cg.rms.dto.Company;
 
 //Repository
-public interface CompanyRepository extends JpaRepository<Company, Long> {
+public interface CompanyRepository extends JpaRepository<Company, Integer> {
 
 	//@Query("SELECT FROM Company c where t.email = :email") 
     public Company findByEmail(@Param("email") String email);
